Rewrite ESP8266WebServer using templates
Refactor the three versions of ESP8266WebServer and ServerSecure to a
single instance of a templated class.  Use typedefs to enable old,
non-templated names to be used (no user changes required).
