+++ libraries/ESP8266WebServer/src/ESP8266WebServerSecure.h
@@ -19,5 +19,13 @@
   Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 */
 
-#include "ESP8266WebServerSecureAxTLS.h"
-#include "ESP8266WebServerSecureBearSSL.h"
+#include "ESP8266WebServer.h"
+
+namespace axTLS {
+  typedef ESP8266WebServerTemplate<WiFiServerSecure, WiFiClientSecure> ESP8266WebServerSecure;
+};
+
+namespace BearSSL {
+  typedef ESP8266WebServerTemplate<WiFiServerSecure, WiFiClientSecure> ESP8266WebServerSecure;
+};
+
