+++ libraries/ESP8266WebServer/src/detail/RequestHandler.h
@@ -1,13 +1,13 @@
 #ifndef REQUESTHANDLER_H
 #define REQUESTHANDLER_H
 
-class RequestHandler {
+template<class ServerClass, class ClientClass> class RequestHandler {
 public:
     virtual ~RequestHandler() { }
     virtual bool canHandle(HTTPMethod method, String uri) { (void) method; (void) uri; return false; }
     virtual bool canUpload(String uri) { (void) uri; return false; }
-    virtual bool handle(ESP8266WebServer& server, HTTPMethod requestMethod, String requestUri) { (void) server; (void) requestMethod; (void) requestUri; return false; }
-    virtual void upload(ESP8266WebServer& server, String requestUri, HTTPUpload& upload) { (void) server; (void) requestUri; (void) upload; }
+    virtual bool handle(ESP8266WebServerTemplate<ServerClass, ClientClass>& server, HTTPMethod requestMethod, String requestUri) { (void) server; (void) requestMethod; (void) requestUri; return false; }
+    virtual void upload(ESP8266WebServerTemplate<ServerClass, ClientClass>& server, String requestUri, HTTPUpload& upload) { (void) server; (void) requestUri; (void) upload; }
 
     RequestHandler* next() { return _next; }
     void next(RequestHandler* r) { _next = r; }
