+++ libraries/ESP8266WebServer/src/detail/RequestHandlersImpl.h
@@ -7,9 +7,9 @@
 
 using namespace mime;
 
-class FunctionRequestHandler : public RequestHandler {
+template<class ServerClass, class ClientClass> class FunctionRequestHandler : public RequestHandler<ServerClass, ClientClass> {
 public:
-    FunctionRequestHandler(ESP8266WebServer::THandlerFunction fn, ESP8266WebServer::THandlerFunction ufn, const String &uri, HTTPMethod method)
+    FunctionRequestHandler(typename ESP8266WebServerTemplate<ServerClass, ClientClass>::THandlerFunction fn, typename ESP8266WebServerTemplate<ServerClass, ClientClass>::THandlerFunction ufn, const String &uri, HTTPMethod method)
     : _fn(fn)
     , _ufn(ufn)
     , _uri(uri)
@@ -34,7 +34,7 @@ public:
         return true;
     }
 
-    bool handle(ESP8266WebServer& server, HTTPMethod requestMethod, String requestUri) override {
+    bool handle(ESP8266WebServerTemplate<ServerClass, ClientClass>& server, HTTPMethod requestMethod, String requestUri) override {
         (void) server;
         if (!canHandle(requestMethod, requestUri))
             return false;
@@ -43,7 +43,7 @@ public:
         return true;
     }
 
-    void upload(ESP8266WebServer& server, String requestUri, HTTPUpload& upload) override {
+    void upload(ESP8266WebServerTemplate<ServerClass, ClientClass>& server, String requestUri, HTTPUpload& upload) override {
         (void) server;
         (void) upload;
         if (canUpload(requestUri))
@@ -51,13 +51,13 @@ public:
     }
 
 protected:
-    ESP8266WebServer::THandlerFunction _fn;
-    ESP8266WebServer::THandlerFunction _ufn;
+    typename ESP8266WebServerTemplate<ServerClass, ClientClass>::THandlerFunction _fn;
+    typename ESP8266WebServerTemplate<ServerClass, ClientClass>::THandlerFunction _ufn;
     String _uri;
     HTTPMethod _method;
 };
 
-class StaticRequestHandler : public RequestHandler {
+template<class ServerClass, class ClientClass> class StaticRequestHandler : public RequestHandler<ServerClass, ClientClass> {
 public:
     StaticRequestHandler(FS& fs, const char* path, const char* uri, const char* cache_header)
     : _fs(fs)
@@ -80,7 +80,7 @@ public:
         return true;
     }
 
-    bool handle(ESP8266WebServer& server, HTTPMethod requestMethod, String requestUri) override {
+    bool handle(ESP8266WebServerTemplate<ServerClass, ClientClass>& server, HTTPMethod requestMethod, String requestUri) override {
         if (!canHandle(requestMethod, requestUri))
             return false;
